Begin work on changing from PID to Bang-Bang
Started work on phasing out PID algorithm for shooters. Replacing PID is
bang-bang, an algorithm that, rather using variable voltage, uses only
full voltage or no voltage to get to the desired value. In this commit,
the Shooter subsystem has been modified to no longer use PID to get to
the required RPM.